// projects/uikit-components/src/lib/helpers/keys.enums.ts
export enum KEY_CODE {
    BACKSPACE = 8,
    TAB = 9,
    ENTER = 13,
    SPACE = 32,
    END = 35,
    HOME = 36,
    DELETE = 46,
    DECIMAL_POINT = 110,
    PERIOD = 190,
    COMMA = 188,
    // Arrows
    LEFT_ARROW = 37,
    UP_ARROW = 38,
    RIGHT_ARROW = 39,
    DOWN_ARROW = 40,
    // Numbers
    NUMBER_0 = 48,
    NUMBER_1 = 49,
    NUMBER_2 = 50,
    NUMBER_3 = 51,
    NUMBER_4 = 52,
    NUMBER_5 = 53,
    NUMBER_6 = 54,
    NUMBER_7 = 55,
    NUMBER_8 = 56,
    NUMBER_9 = 57,
    DIGIT_0 = 96,
    DIGIT_1 = 97,
    DIGIT_2 = 98,
    DIGIT_3 = 99,
    DIGIT_4 = 100,
    DIGIT_5 = 101,
    DIGIT_6 = 102,
    DIGIT_7 = 103,
    DIGIT_8 = 104,
    DIGIT_9 = 105,
    // Letters
    LETTER_A = 65,
    LETTER_C = 67,
    LETTER_V = 86,
    LETTER_X = 88,
    LETTER_Y = 89,
    LETTER_Z = 90,
  }